refactor: 공통적으로 쓰이는 validator 분리
